updated my test run bash script I still have it in me :D I did have to look up how to check if a file exists though

+++ TestRunScript.bash
@@ -1,5 +1,12 @@
 #!/usr/bin/env bash
-cd ./preprocessor/
-gcc _putchar.c main.c *-*.h -o testRun -Wall -Werror -Wextra -pedantic -std=gnu89
+cd ./doubly_linked_lists/ || exit
+gcc main.c *-*.c -o testRun -Wall -Werror -Wextra -pedantic -std=gnu89
+if ! test -f ./testRun
+then
+  echo "Couldn't compile."
+  exit
+fi
+echo
 ./testRun
+echo
 rm testRun
